zoomit/requirements.txt
djangorestframework_simplejwt==4.6.0
Django==3.1.2
djangorestframework==3.12.2
